GameServer/web_server/srcs/tasks/requirements.txt
PyYAML==3.12
supervisor==3.3.3